+++ src/main/java/com/atlauncher/data/minecraft/ArgumentRule.java
@@ -18,7 +18,6 @@
 package com.atlauncher.data.minecraft;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 import com.atlauncher.annot.Json;