+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add in servers tab to manage servers created through the launcher
 - Add Curse packs from the packs tab
 - [#370] fix JVM arguments not adding in correct order
+- Fix adding mods showing modpacks and add dropdown for resource packs